requirements.txt
# Data Collection & Processing
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
numpy>=1.24.0

# NLP & Text Processing
nltk>=3.8.0
tiktoken>=0.5.0

# Embeddings & Vector Store
sentence-transformers>=2.2.0
chromadb>=0.4.0

# BM25 Search
rank-bm25>=0.2.2

# Re-ranking
torch>=2.0.0
transformers>=4.35.0

# ONNX Optimization (2-3x CPU speedup)
onnxruntime>=1.16.0
optimum[onnxruntime,exporters]>=1.14.0

# LLM Integration
ollama>=0.1.0
groq>=0.4.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Frontend & Dashboard
streamlit>=1.28.0

# Evaluation
ragas>=0.1.0
datasets>=2.14.0

# Observability
langfuse>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
uuid6>=2024.1.12

# Notebook
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
